Move language documentation to the right place
Discussion of what macros are was in
implementation.md, and had no discussion of their
implementation — it was language reference
material. Move it to language.md, and move the
code to scripts/macros.fe and update the tests. In
the process, discover and fix a bug in the example
code (`(while (< i 0))` instead of `(while (< i
10))`).